+++ Renderer/Renderer/Model/MeshForCPU.h
@@ -15,13 +15,13 @@ namespace NS
 	struct MeshForCPU
 	{
 		std::vector<Vertex> vertices;
-		std::vector<uint16_t> indices;
+		std::vector<uint32_t> indices;
 		std::string albedoTextureFilename;
 	};
 
 	struct MeshNormalForCPU
 	{
 		std::vector<VertexForDrawingNormal> vertices;
-		std::vector<uint16_t> indices;
+		std::vector<uint32_t> indices;
 	};
 }
